SCALE-SEAM (88) — the operate-phase cluster, first slice out of the unmapped 126
Nine methods out of client.ts (1,015 -> 953) into a new
apps/web/src/api/operations.ts. What they answer together: the building is
built and running. Maintenance (cmmsGeneratePm, cmmsKpis), consumption
(energyActual, energyBenchmarkStatus, esgSummary), condition and capital
(fcaIndex, fcaPortfolio, reserveStudy), plus twinReadiness as the data
precondition for all three.

CHECKED AGAINST THE BACKEND RATHER THAN ASSUMED: all nine are served by
aec_api/routers/operations.py, whose own docstring names the same cluster.
That is corroboration from a source this extraction did not write, which is
the point after (87) — a completeness claim computed over a self-authored
list is confident and unfounded.

THREE DID NOT COME, AND THE WORDS SAY OTHERWISE ON EVERY ONE.

lifecycle / lifecycleSeed are not asset lifecycle. Their banner reads
"design lifecycle (RIBA/AIA phases + itemized soft costs)" and the payload
is design_fee_pct, deliverables, iso_status, soft_costs — design stage gates
and what the design costs, years before anything operates. Fourth shared-word
trap in this sequence, after entitlements, view and carbon.

projectCarbon is EMBODIED carbon, a design-phase estimate; esgSummary's GHG
comes from metered utility data. Same molecule, opposite ends of the asset's
life.

camReconciliation shared reserveStudy's banner AND its backend router and
still stayed. It allocates the recoverable pool across TENANTS by share and
returns balance_due per suite: a lease-revenue answer, not a
building-condition one. It waits for a rent-roll slice with rentRollScrub,
netEffectiveRent and normalizeT12. Moving it along to empty a banner would
have been the route-and-neighbour reasoning this sequence keeps rejecting,
so the banner it sat under is rewritten rather than left naming a method
that has gone.

AND (87)'s OWN BANNER HAD ALREADY DRIFTED. The STAYING banner listed
esgSummary among its examples of work still above the line — offering, as
evidence of what is unfinished, a method this slice moved. Corrected to 117,
re-pointed at camReconciliation, and it now instructs the reader to
re-derive the number and names from the file rather than read them there.
The count itself is asserted by test_roadmap_status.py.

117 methods remain above the banner and there is still no map for them.
This was one cluster read out of the file, not a plan for the rest.

Verified: tsc exit 0, eslint exit 0, 203 files / 2,059 web tests, ruff clean
over the whole tree, and all seven doc and size gates OK. The ratchet's
history-chain gate added in (87) validated its own new (1015 -> 953) entry.
